refactor: 将mapOperation移动到单独的文件中
