+++ Lab04_4.csproj
@@ -277,6 +277,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Form_4.MapOperation.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Form_4.cs">
       <SubType>Form</SubType>
     </Compile>
